Title as a first-class citizen of a post.
Usually when I start wrtiting I have a title in mind already and put it
as a top-level headline in markdown. Then I copy-paste contents to new
post template — always moving top-level headline into title meta.

This title is then rendered as h1 (#) and gets special treatment via CSS.
Yet this often leaves doubt how should other headlines be structured.

It became obvious that it is impossible to expect from others to
remember or know about starting from h2 (##). So it seems rather obvious
to move title where it belongs and move the dull task of extracting it
to the software, sparing some human brain cycles for making typos elsewhere.

This change should be followed in blog engine soon.