Update 2021-10-23-less-known-capability-of-rubys-json-dot-parse.md
I tried to play around with the above code both in a IRB session and in a simple *.rb file to check it out. It seems like you should always require 'json' on the top of the file except if you use `-r` option when running it from the command line. The same comes for an IRB session.