Optimize query
`User.where.not(category: "")` will also find records where column is not null or empty (`SELECT "users".* FROM "users" WHERE "users"."category" != ?  [["category", ""]`) but in a more efficient way than using the array form (`NOT IN`). Because of the nature of `not`, it does not fetch records where `category` is `null`. You can read more about this topic in https://thoughtbot.com/blog/activerecord-s-where-not-and-nil